fixing seeders

+++ backend/db/migrations/20220527221247-create-review.js
@@ -30,7 +30,7 @@ module.exports = {
       },
       stars: {
         allowNull: false,
-        type: Sequelize.INTEGER(5)
+        type: Sequelize.INTEGER
       },
       createdAt: {
         allowNull: false,
